perf: build url regex in advance
